Add tidied and avg data transforms, adel read logic
Added the transformation of the tidied dataset by using the melt
function, which made it easy to create the dataset of averages for the
variables.

Added check for file existence when scripts starts. If dataset
directory does not exists then unzip the data. If zip file does not
exist then stop execution with an error.